move towards module folders.  Break up top-level modules.
