+++ .travis.yml
@@ -4,7 +4,6 @@ python:
   - "2.7"
   #- "3.4" # comment python 3 for now (a few upstream libs don't support it.)
 install:
-  # TODO: Use a "latest" url
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
       wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh;
     else
