--- topik/intermediaries/corpus_dictionary.py
@@ -1,2 +1,0 @@
-import gensim
-# TODO: implement
