+++ topik/tests/test_readers.py
@@ -1,11 +1,11 @@
 import unittest
 
 import nose.tools as nt
-import elasticsearch
 
-from topik.readers import read_input
-import topik.readers as readers
-from topik.intermediaries.raw_data import ElasticSearchCorpus
+import elasticsearch
+from topik.fileio.readers import read_input
+import topik.fileio.readers as readers
+from topik.fileio.raw_corpus import ElasticSearchCorpus
 from topik.tests import test_data_path
 
 INDEX = "test_elastic"
