+++ topik/tests/test_vectorizer.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from topik.readers import read_input
+from topik.fileio.readers import read_input
 from topik.tests import test_data_path
